backend/requirements.txt
# FastAPI & Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# MongoDB
motor>=3.3.2
pymongo>=4.6.1

# LangChain & AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-mongodb>=0.1.0
langgraph>=0.0.20

# Embeddings
openai>=1.10.0
sentence-transformers>=2.3.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
unstructured>=0.12.0

# Vector Store
chromadb>=0.4.22

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
tenacity>=8.2.3

# Build & Obfuscation (optional)
pyinstaller>=6.3.0
nuitka>=1.9.0
